+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 - Added SECURITY.md
 - Removed extra order text and added missing y
+- Hide 1-click setup button when setup
+- Added Apple Pay, Google Pay and Klarna as payment methods
+- Remove unused isAutoCapture helper 
 
 ## [1.0.4] - 2024-10-16
 - Added general platform that identifies the plugin with the API.